Added authentication to delete

+++ back/scripts/routers/removeEnd.js
@@ -6,6 +6,8 @@ const router = express.Router();
 router.delete('', async (req, res, next) => {
   try {
     const id = req.body.id;
+    const auth = req.headers.auth;
+    if (auth !== process.env.PASSWORD) throw errorCodes.badAuth;
     if (!id) throw errorCodes.requestInputInvalid;
     const mongoRes = await questionQueries.deleteQuestionById(id);
     res.json({ message: mongoRes });
